move unit test to own file
